requirements.txt
fastapi
uvicorn
pydantic
numpy
onnxruntime
transformers
huggingface-hub
prometheus-fastapi-instrumentator
python-logstash
psutil